backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
aliyun-python-sdk-sts==3.1.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
backcall==0.2.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==6.1.0
celery==5.5.0
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cron-descriptor==1.4.5
cryptography==41.0.5
daphne==4.1.2
dataclasses-json==0.6.7
dateutils==0.6.12
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.20
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docopt==0.6.2
eventlet==0.40.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
frozenlist==1.5.0
gevent==25.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.1.0
ipython==8.12.3
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
langchain==0.3.26
langchain-openai==0.3.28
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.1.3
msgpack==1.1.0
multidict==6.3.2
mypy_extensions==1.1.0
narwhals==1.45.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.0.2
openai==1.93.0
orjson==3.10.18
oss2==2.19.1
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pickleshare==0.7.5
pika==1.3.2
pillow==11.1.0
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.6.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.31.1
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
pytest==8.4.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-decouple==3.8
pytz==2025.1
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
service-identity==24.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.46.1
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.19.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
